+++ projects/ngx-stripe-tester/src/app/app.module.ts
@@ -1,7 +1,9 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
+import { ReactiveFormsModule } from '@angular/forms';
 import { ServiceWorkerModule } from '@angular/service-worker';
 
 import { NgxStripeModule } from 'ngx-stripe';
@@ -9,7 +11,8 @@ import { NgxStripeModule } from 'ngx-stripe';
 import { environment } from '../environments/environment';
 
 import { SharedModule } from './shared/shared.module';
-import { CoreModule } from './core/core.module';
+
+import { CoreModule, PLUTO_ID } from './core';
 
 import { AppComponent } from './app.component';
 import { ROUTES } from './app.routing';
@@ -18,8 +21,10 @@ import { ROUTES } from './app.routing';
   declarations: [AppComponent],
   imports: [
     BrowserModule,
+    HttpClientModule,
     RouterModule.forRoot(ROUTES, { relativeLinkResolution: 'legacy' }),
-    NgxStripeModule.forRoot('pk_test_nDR7IWEIGLp4a1SBtqKH5eyg'),
+    NgxStripeModule.forRoot('pk_test_51Ii5RpH2XTJohkGafOSn3aoFFDjfCE4G9jmW48Byd8OS0u2707YHusT5PojHOwWAys9HbvNylw7qDk0KkMZomdG600TJYNYj20'),
+    ReactiveFormsModule,
     NoopAnimationsModule,
     ServiceWorkerModule.register('ngsw-worker.js', {
       enabled: environment.production
@@ -27,7 +32,12 @@ import { ROUTES } from './app.routing';
     SharedModule,
     CoreModule
   ],
-  providers: [],
+  providers: [
+    {
+      provide: PLUTO_ID,
+      useValue: '449f8516-791a-49ab-a09d-50f79a0678b6'
+    }
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule {}