+++ projects/ngx-stripe-tester/src/app/core/index.ts
@@ -1,0 +1,4 @@
+export * from './pluto/client-id.provider';
+export * from './pluto/pluto.service';
+
+export * from './core.module';