+++ projects/ngx-stripe-tester/src/app/core/pluto/client-id.provider.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from '@angular/core';
+
+export const PLUTO_ID = new InjectionToken<string>('[PLUTO] ClientID');