+++ projects/ngx-stripe-tester/src/app/shared/material.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 
 import { MatDividerModule } from '@angular/material/divider';
 import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
 import { MatListModule } from '@angular/material/list';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
 import { MatSidenavModule } from '@angular/material/sidenav';
@@ -10,6 +11,7 @@ import { MatToolbarModule } from '@angular/material/toolbar';
 const modules = [
   MatDividerModule,
   MatIconModule,
+  MatInputModule,
   MatListModule,
   MatProgressBarModule,
   MatSidenavModule,